refactor: Announcements.jsx - 반복되는 코드 renderInfoItem로 추출
